Web design and clear all warnings

+++ client/src/components/home/Navbar.js
@@ -2,6 +2,7 @@ import React, { useEffect, useState } from 'react';
 import { Link, useNavigate } from 'react-router-dom';
 import { resetLogin } from '../../actions/userAction'
 import { useDispatch, useSelector } from 'react-redux';
+import Swal from 'sweetalert2';
 
 const Navbar = () => {
   const [token, setToken] = useState(null);
@@ -17,9 +18,26 @@ const Navbar = () => {
   
   const navigate = useNavigate()
   const logoutHandler = () => {
-    localStorage.clear();
-    dispatch(resetLogin())
-    navigate('/user/login');
+    Swal.fire({
+      title: 'Are you sure?',
+      text: "You want to logout?",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes'
+  }).then((result) => {
+      if (result.isConfirmed) {
+        localStorage.clear();
+        dispatch(resetLogin())
+        navigate('/user/login');
+        // Swal.fire(
+        //     'Retrieved!',
+        //     'Your post has been successfully retrieved.',
+        //     'success'
+        // )
+      }
+    })
   }
 
   return (
